statistics

+++ src/main/java/com/shulgin/yandex_autumn/service/ElementService.java
@@ -1,23 +1,25 @@
 package com.shulgin.yandex_autumn.service;
 
 import com.shulgin.yandex_autumn.entity.Element;
-import com.shulgin.yandex_autumn.entity.ElementType;
 import com.shulgin.yandex_autumn.entity.File;
 import com.shulgin.yandex_autumn.entity.Folder;
-import com.shulgin.yandex_autumn.exception.ValidationException;
+import com.shulgin.yandex_autumn.entity.Statistic;
 import com.shulgin.yandex_autumn.repository.ElementRepository;
+import com.shulgin.yandex_autumn.repository.StatisticRepository;
 import org.springframework.stereotype.Service;
 
-import java.time.ZonedDateTime;
 import java.util.Optional;
 import java.util.UUID;
 
 @Service
 public class ElementService {
     private final ElementRepository elementRepository;
+    private final StatisticRepository statisticRepository;
 
-    public ElementService(ElementRepository elementRepository) {
+    public ElementService(ElementRepository elementRepository,
+                          StatisticRepository statisticRepository) {
         this.elementRepository = elementRepository;
+        this.statisticRepository = statisticRepository;
     }
 
     public void saveFolder(Folder folder){
@@ -25,6 +27,7 @@ public class ElementService {
     }
 
     public void saveFile(File file) {
+        statistic(file, file.getSize());
         save(file);
         updateSize(file, file.getSize());
     }
@@ -46,6 +49,7 @@ public class ElementService {
             if(optionalParentElement.isPresent())
             {
                 Element parentElement = optionalParentElement.get();
+                statistic(parentElement, parentElement.getSize() + size);
                 parentElement.setSize(parentElement.getSize() + size);
                 save(parentElement);
                 updateSize(parentElement, size);
@@ -68,6 +72,29 @@ public class ElementService {
     }
 
     public void deleteById(UUID id) {
-        elementRepository.deleteById(id);
+        Optional<Element> optionalElement = elementRepository.findElementById(id);
+        if(optionalElement.isPresent()) {
+            Element element = optionalElement.get();
+            long size = element.getSize();
+            updateSize(element, -size);
+            elementRepository.deleteById(id);
+        }
+    }
+
+    public void statistic(Element element, long size) {
+        UUID id = element.getId();
+        Optional<Element> oldOptionalElement = elementRepository.findElementById(id);
+        long oldSize = -1;
+        if(oldOptionalElement.isPresent()) {
+            Element oldElement = oldOptionalElement.get();
+            oldSize = oldElement.getSize();
+        }
+        if(oldSize != size) {
+            Statistic statistic = new Statistic();
+            statistic.setDate(element.getDate());
+            statistic.setSize(size);
+            statistic.setElemId(element.getId());
+            statisticRepository.save(statistic);
+        }
     }
 }
